Rework connection unit tests
A unique_ptr was being used, when there is no need.
Setup and TearDown of the connection is removed to be
done always, because it causes a lot of pain in every test.

A test case is removed because there is no guarantee of the
on_error callback to be called when a connection cannot be
established. For that reason a new function is also added
to check for any status for a given time.

This one will be used in the future everywhere, but a bigger
refactor is needed.