Add client tests
By using mocks, the client can now be tested against a real
server. As it's async, the strategy is to identify the last action
performed by the client, and set the future status to ready
when such call is done.

This commit adds a basic passing test, and more will be
added afterwards.