Fix unstabilities in connection test
This commit changes expectations for assertions, since they are
actually very needed. Moreover, and unfortunately, a sleep is added.

This test is unstable due to, despite the efforts on checking the
status,  the connection is not ready. Closing a connection which is
not fully open ends up in the nghttp2 client in the following error:
Assertion failed: !inside_callback_ (asio_client_session_impl.cc: enter_callback: 611)