Add tests for TLS connections
To be able to run this test some cryptographic material is needed:
There should be a trusted CA in the development container, this CA
will be in charge of signing the certificate that will be used to
start a server using TLS.

All that functionality is provided by the "generateCryptographicMaterial"
script. The development Dockerfile is in charge of calling it so the certs
are generated and the tlsConnection_test tests can be run.