Complete https contribution
In this commit the work from pr#50 is completed, in order to
speed the revision process up.

The Dockerfile is kept as it was, so hermes image build is not
executing the tests, or executing scripts. Instead of that, the
script is moved (and name changed to follow this repo's nomenclature)
to its own folder (scripts).

Such script is now executed in the CMake configure step. In this way,
tests won't fail and it's always executed when preparing for compilation.
Thus, no need to generate new versions of hermes_base (that would break
the pipeline).

Tests for new boolean in the script added to enforce coverage.
Tests for the connection changed as parameterized to avoid repeating code.
New tests for secure client added following the same approach.