Change kind action in CI
The server version for the engineerd takes too long and
installs an old server that the client cannot talk to.
Trying a new one, setting the version of kubectl and also
hoping for it to take less than 5 mins. In their doc it
seems that the default is 1 min, so maybe that's a good sign.

Using a commit id since the fix hasn't been released yet.