hermes_base:0.0.2 - Bump up boost and nghttp versions
Boost bumped up to 1.76.0 from 1.67.0. Version 1.77.0 not used
because it makes the server crash every time the object is deleted.
This makes the unit test to not work.

A warning appeared for boost/bind, in which it's mandatory now
to use boost/bind/bind.hpp.

Nghttp2 version bumped up to 1.45.1, which is the latest.

Docker image for hermes_base generated and pushed manually, so all
documentation updated accordingly.