Parse variables and ranges in headers
This commit adds the functionality of substituting the
variables and ranges (as in body and url) for
headers. The mechanism is the same. The code is not the finest I've done
but I think it will be the same using other algos.

If this is a problem, headers can be parsed as a json string and work from
there.