Allow saving headers from answer
This commit kind of undos all the previous work in this PR.
It is not easy to find a good, general way to handle
variables, ranges, and message modifiers
at the same time. So in the end, let's go step by step.

save_from_answer API is now broken, so headers are added.
add_from_save_to_body too.

"name" field is removed from both structures, so now the name is
taken from the key of the newly formed json object, which
now contains path and value_type, as before.