Fix header post processing
Headers were not being substituted in the next message as parsed,
but only on "add from saved to body".
A wrong test was hindering this issue.

This commit fixes that by making the test correct (in the first
place) and forcing variable parsing post
process.

Now a first post processing round is done when building a new script, and a
second one when getting the new message.
In this way is a bit more intuitive to use (hopefully).