Fix message modifications for empty body
Json parsing was returning an error when the body is empty.
This is left like that on purpose to avoid future errors. Instead, a
(not very performing) check is added, and when the body is empty, a tmp one
with an empty body is created before trying to add saved values.